+++ packages/state_forge_core/lib/state_forge_core.dart
@@ -4,6 +4,7 @@ library;
 export 'src/async_state.dart';
 export 'src/config.dart';
 export 'src/forge_observer.dart';
+export 'src/effect_store.dart';
 export 'src/store.dart';
 export 'src/mixins/composited_store.dart';
 export 'src/mixins/persistable_store.dart';